web-server/pyproject.toml
[tool.poetry]
name = "volunteer_toronto_be"
version = "0.1.0"
description = ""
authors = ["Erik Derohanian <erik@erik-j.de>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.4"
djangorestframework = "^3.12.2"
black = "^20.8b1"
drf-yasg = "^1.20.0"
httpie = "^2.4.0"

[tool.poetry.dev-dependencies]
pylint = "^2.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
